OPAL-1915: Moving User, Group and Status from org.obiba.opal.core to org.obiba.opal.core.domain
