+++ opal-core/src/main/java/org/obiba/opal/core/domain/user/User.java
@@ -7,7 +7,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.obiba.opal.core.user;
+package org.obiba.opal.core.domain.user;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
