+++ opal-core/src/main/java/org/obiba/opal/core/service/impl/UserService.java
@@ -11,8 +11,8 @@ package org.obiba.opal.core.service.impl;
 
 import java.util.List;
 
-import org.obiba.opal.core.user.Group;
-import org.obiba.opal.core.user.User;
+import org.obiba.opal.core.domain.user.Group;
+import org.obiba.opal.core.domain.user.User;
 
 @SuppressWarnings("UnusedDeclaration")
 public interface UserService {
